+++ app/src/views/CountryNsOverviewActivities/NsDirectoryInitiatives/InitiativeCard/styles.module.css
@@ -1,0 +1,38 @@
+.initiative-card {
+    border-radius: var(--go-ui-border-radius-lg);
+    box-shadow: var(--go-ui-box-shadow-md);
+    padding: var(--go-ui-spacing-lg);
+
+    .figures {
+        display: flex;
+        justify-content: space-between;
+
+        .vertical_separator {
+            flex-shrink: 0;
+            border-left: var(--go-ui-width-separator-sm) solid var(--go-ui-color-separator);
+        }
+
+    }
+
+    .year_allocation_section {
+        align-self: center;
+
+        .year_allocation_value {
+            font-weight: var(--go-ui-font-weight-bold);
+        }
+
+        .year_allocation_label {
+            color: var(--go-ui-color-text-light);
+        }
+    }
+
+    .footer-content {
+        flex-grow: 1;
+
+        .separator {
+            flex-shrink: 0;
+            margin-bottom: var(--go-ui-spacing-sm);
+            border-bottom: var(--go-ui-width-separator-thin) solid var(--go-ui-color-separator)
+        }
+    }
+}
