+++ app/src/views/CountryNsOverviewActivities/NsDirectoryInitiatives/i18n.json
@@ -1,7 +1,7 @@
 {
     "namespace": "nsDirectoryInitiatives",
     "strings": {
-        "nSDirectoryInitiativesTitle": "NSD Initiatives",
+        "nSDirectoryInitiativesTitle": "Active NSD funded initiatives",
         "nSDirectoryInitiativesDescription": "The following data displays ongoing and implemented/closed Initiatives funded  by the different NSD funding mechanisms within the country, which are reported by the respective fund office. Contacts: CBF Office, NSIA Office, ESF Secretariat.",
         "source": "Source",
         "ifrc": "IFRC"
