+++ app/src/views/CountryNsOverviewActivities/index.tsx
@@ -48,6 +48,7 @@ import { useRequest } from '#utils/restRequest';
 
 import Filter, { FilterValue } from './Filters';
 import Map from './Map';
+import NationalSocietyDirectoryInitiatives from './NsDirectoryInitiatives';
 
 import i18n from './i18n.json';
 import styles from './styles.module.css';
@@ -524,6 +525,7 @@ export function Component() {
                     keySelector={numericIdSelector}
                 />
             </Container>
+            <NationalSocietyDirectoryInitiatives />
         </Container>
     );
 }
