--- app/src/views/CountryNsOverviewContextAndStructure/NsDirectoryInitiatives/InitiativeCard/i18n.json
@@ -1,12 +1,0 @@
-{
-    "namespace": "initiativeCard",
-    "strings": {
-        "initiativeFundNameTitle": "Fund Name",
-        "initiativeYearApprovedTitle": "Year Approved",
-        "initiativeTitle": "Initiative Title",
-        "initiativeCategoriesTitle": "Categories",
-        "initiativeAllocationTitle": "Allocation CHF",
-        "initiativeFundingPeriodTitle": "Funding Period",
-        "initiativeMonthsSuffix": "months"
-    }
-}
