--- app/src/views/CountryNsOverviewContextAndStructure/NsDirectoryInitiatives/InitiativeCard/styles.module.css
@@ -1,21 +1,0 @@
-.initiative-card {
-    border-radius: var(--go-ui-border-radius-lg);
-    box-shadow: var(--go-ui-box-shadow-md);
-    padding: var(--go-ui-spacing-lg);
-
-    .figures {
-        display: flex;
-        flex-direction: column;
-        gap: var(--go-ui-spacing-xs);
-
-    }
-    .footer-content {
-        flex-grow: 1;
-
-        .separator {
-            flex-shrink: 0;
-            margin-bottom: var(--go-ui-spacing-sm);
-            border-bottom: var(--go-ui-width-separator-thin) solid var(--go-ui-color-separator)
-        }
-    }
-}
